Manipulating Bits
Manipulating Bits These lines of code were produced as part of course INF559 (Computer and Operating System Architecture (2019-2020)) on bit manipulation using binary operators c with restrictions. There are 13 exercices manipulation of bits that have been given by the teachers and for each exercise, there are properties to respect in the comments. The basic code comes from the teacher, but the exercises were done by ourselves.